pyproject.toml
[tool.poetry]
name = "bsort"
version = "0.1.0"
description = "CLI tool for Adamata Take Home Test YOLO model"
authors = ["Brillyando M.A."]
readme = "README.md"
packages = [{include = "bsort"}]

[tool.poetry.dependencies]
python = ">=3.10"
click = "^8.3.1"
pyyaml = "^6.0.3"
ultralytics = "^8.3.229"
opencv-python = "^4.12.0.88"
albumentations = "^2.0.8"
tqdm = "^4.67.1"
matplotlib = "^3.10.7"
wandb = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
pylint = "^3.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
bsort = "bsort.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pylint.messages_control]
disable = [
    "C0111",
    "C0103",
    "R0913",
]
max-line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=bsort --cov-report=term-missing"
